Edited 26

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -32,13 +32,13 @@ long double k2;
 		{
 
 
-			h = k2 + k1;
+			(long) h = k2 + k1;
 			
 			k1 = 0;
-			k1 = k2;
+			(long) k1 = k2;
 			
 			k2 = 0;
-			k2 = h;
+			(long) k2 = h;
 			
 			printf("%Lf", h);
 			if (f != 97)
